# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CarCraftr API",
    "description": "A comprehensive car information and community platform API",
    "version": "1.0.0",
    "contact": {
      "name": "CarCraftr Team",
      "email": "support@carcraftr.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    },
    {
      "url": "https://api.carcraftr.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/market/car/{carId}": {
      "get": {
        "summary": "Get market data for car",
        "description": "Retrieve market data for a specific car",
        "parameters": [
          {
            "name": "carId",
            "in": "path",
            "required": true,
            "description": "Car ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Market data for the car",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketData"
                }
              }
            }
          }
        }
      }
    },
    "/api/market/car/{carId}/price-trends": {
      "get": {
        "summary": "Get price trends",
        "description": "Retrieve price trends for a specific car",
        "parameters": [
          {
            "name": "carId",
            "in": "path",
            "required": true,
            "description": "Car ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Price trends data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "date": {
                        "type": "string",
                        "format": "date"
                      },
                      "price": {
                        "type": "number"
                      },
                      "trend": {
                        "type": "string",
                        "enum": ["up", "down", "stable"]
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/reviews/car/{carId}": {
      "get": {
        "summary": "Get reviews for car",
        "description": "Retrieve all reviews for a specific car",
        "parameters": [
          {
            "name": "carId",
            "in": "path",
            "required": true,
            "description": "Car ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of reviews",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Review"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/reviews": {
      "post": {
        "summary": "Create review",
        "description": "Add a new review for a car",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["carId", "rating", "title", "content"],
                "properties": {
                  "carId": {
                    "type": "string"
                  },
                  "rating": {
                    "type": "number",
                    "minimum": 1,
                    "maximum": 5
                  },
                  "title": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  },
                  "pros": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "cons": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Review created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Review"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MarketData": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "carId": {
            "type": "string"
          },
          "priceTrends": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "date": {
                  "type": "string",
                  "format": "date"
                },
                "price": {
                  "type": "number"
                },
                "trend": {
                  "type": "string",
                  "enum": ["up", "down", "stable"]
                }
              }
            }
          },
          "resaleValues": {
            "type": "object",
            "properties": {
              "oneYear": {
                "type": "number"
              },
              "threeYear": {
                "type": "number"
              },
              "fiveYear": {
                "type": "number"
              }
            }
          },
          "marketDemand": {
            "type": "object",
            "properties": {
              "demandLevel": {
                "type": "string",
                "enum": ["low", "medium", "high"]
              },
              "supplyLevel": {
                "type": "string",
                "enum": ["low", "medium", "high"]
              },
              "marketScore": {
                "type": "number",
                "minimum": 0,
                "maximum": 100
              }
            }
          },
          "bestTimeToBuy": {
            "type": "object",
            "properties": {
              "season": {
                "type": "string",
                "example": "winter"
              },
              "reasoning": {
                "type": "string"
              },
              "discountPercentage": {
                "type": "number"
              }
            }
          },
          "dealerInventory": {
            "type": "object",
            "properties": {
              "availableUnits": {
                "type": "number"
              },
              "averagePrice": {
                "type": "number"
              },
              "priceRange": {
                "type": "object",
                "properties": {
                  "min": {
                    "type": "number"
                  },
                  "max": {
                    "type": "number"
                  }
                }
              }
            }
          },
          "marketInsights": {
            "type": "object",
            "properties": {
              "lastUpdated": {
                "type": "string",
                "format": "date-time"
              },
              "insights": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "Review": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "carId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "rating": {
            "type": "number",
            "minimum": 1,
            "maximum": 5
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "pros": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "cons": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "likes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "helpfulCount": {
            "type": "number"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
} 